feat(messages): :sparkles: add mailing at admin panel [4/7]
Добавлена возможность начала рассылки в админ-панель
