+++ handlers/commands.py
@@ -5,21 +5,18 @@ from asyncio import sleep
 from logging import error
 from random import uniform
 
-from aiogram import Bot, Router
+from aiogram import Router
 from aiogram.filters import Command, CommandStart
 from aiogram.fsm.context import FSMContext
 from aiogram.types import Message
-from aiogram.utils.text_decorations import html_decoration
 
 from config_reader import config
 from database.core import db
-from database.currencies import calculate_precise_growth_chance, change_coin
-from database.messages import send_broadcast_message
+from database.currencies import calculate_precise_growth_chance
 from database.user import check_casino_balance, check_profile
 from keyboards.builders import create_main_buttons
 from keyboards.reply import register
 from states.enums import UserStatus
-from states.fsm_states import BroadcastText
 
 router = Router()
 
@@ -53,59 +50,12 @@ async def check_handler(message: Message, state: FSMContext):
     random_nu = uniform(2.50, 5.00)
 
     await message.answer(f"Текущий статус: {status}\n\nДанные Interaction: {data}\n\nRandom: {random_nu}", parse_mode=None)
-    
-
-@router.message(Command("change"))
-async def change_handler(message: Message, bot: Bot):
-    await change_coin("st", bot)
-    await message.answer("Валюта ST изменена\n\n")
 
 
 @router.message(Command("chance"))
 async def chance_handler(message: Message):
     percent = await calculate_precise_growth_chance("st")
     await message.answer(f"Шанс повышения ST: {percent}")
-
-
-@router.message(Command("users"))
-async def users_handler(message: Message):
-    text = ""
-
-    users = await db.select_data("users", "*", fetch_all=True)
-
-    for i, user in enumerate(users):
-        username = user.get("username")
-        user_id = user.get("id")
-        text += f"{i + 1}. @{username} ({user_id})\n"
-    
-    text += f"\nОбщее кол-во пользователей: {len(users)}"
-    await message.answer(text)
-
-
-@router.message(Command("send"))
-async def send_text_handler(message: Message, state: FSMContext):
-    await message.answer("<b>📝 В следующем сообщении отправьте текст для рассылки</b>")
-    await state.set_state(BroadcastText.sending_text)
-
-
-@router.message(BroadcastText.sending_text)
-async def interaction_amount_handler(message: Message, state: FSMContext, bot: Bot):
-    if message.text:
-        formatted_text = message.text
-    elif message.caption:
-        formatted_text = message.caption
-    else:
-        await message.answer("❌ Пожалуйста, отправьте текст для рассылки")
-        return
-    
-    # Если есть entities, конвертируем их в HTML
-    if message.text and message.entities:
-        formatted_text = html_decoration.unparse(message.text, message.entities)
-    elif message.caption and message.caption_entities:
-        formatted_text = html_decoration.unparse(message.caption, message.caption_entities)
-    
-    await state.update_data(sending_text=formatted_text)
-    await send_broadcast_message(state, bot, message.from_user.id)
 
 
 @router.message(Command("game"))
